requirements.txt
HEAD
python-telegram-bot==20.8
python-telegram-bot==22.3
c5379dd (Mise à jour)
python-dotenv
requests
